More BHAV editor changes:
Sort should be thoroughly fixed.  And quite nice.  IIDSSM.

Highlighting should be more consistent now and I might have fixed a bug (or I might just have not introduced one!).

Did a change for not relinking to ERROR when you delete an instruction that's linked to - highlighted links to missing instructions.

I've changed the label on View Global button.